Fleshed out AI script.
