--- Assets/Block.cs
@@ -1,3 +1,0 @@
-﻿public class Block
-{
-}
