--- Assets/Computer.cs
@@ -1,35 +1,0 @@
-﻿public class Computer : IActor
-{
-    public Inventory inventory;
-    public ActorState state;
-
-    public void MoveCamera()
-    {
-
-    }
-    public void PlaceBlock()
-    {
-
-    }
-    public void EndPlacement()
-    {
-
-    }
-
-    public void Shoot()
-    {
-
-    }
-    public void ShowInventory()
-    {
-
-    }
-    public void Die()
-    {
-
-    }
-
-    public bool IsWinner() {
-        return false;
-    }
-}
