--- Assets/Inventory.cs
@@ -1,3 +1,0 @@
-﻿public class Inventory
-{
-}
