+++ Assets/Scripts/Computer.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 6452b9a484e85ac4583e21d49a97279b
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
