+++ Assets/Scripts/IActor.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: c3e07bad35d4ed741943f3687df2e7f5
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
