+++ Assets/Scripts/Inventory.cs
@@ -1,0 +1,7 @@
+﻿using System.Collections.Generic;
+
+public class Inventory
+{
+    public List<Block> buildingBlocks;
+    public List<Block> shootingBlocks;
+}
