+++ Assets/Scripts/World.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 1a755889a8c16224f8598a6e1a0e85a2
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
